some cosmetic code changes GameList is GameList<T> now ObjectHandler renamed to ObjectReferencesHandler

+++ Airplane/Airplane/DenseObjectsList.cs
@@ -143,7 +143,7 @@ namespace Airplane
         {
             denseobjectlist.Remove((DenseObject)obj);
         }
-        public double Count()
+        public long Count()
         {
             return denseobjectlist.Count;
         }
